few things
- time freeze power up
- controller support
