--- Assets/Powerup_TimeFreeze.cs
@@ -1,5 +1,0 @@
-using UnityEngine;
-
-public class Powerup_TimeFreeze : Collectible
-{
-}
