--- Assets/Powerup_TimeFreeze.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a25866253637d424d850e6db7575629b
