+++ Assets/_Project/Images/Collectibles.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 44d89d28fc3c6454dbac9c1f1f7315ec
-PrefabImporter:
+guid: 0cd1c7c0435e1364eb7ef1d2a69bf7b1
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
