+++ Assets/_Project/Images/Collectibles/collectible_icons.xcf.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 44d89d28fc3c6454dbac9c1f1f7315ec
-PrefabImporter:
+guid: 9eed47eb5feb5e948ad0adba85781582
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
