+++ Assets/_Project/Images/Logos.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 44d89d28fc3c6454dbac9c1f1f7315ec
-PrefabImporter:
+guid: 6c0cb813c3ccb054ba21eb563c1caafc
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
