+++ Assets/_Project/Images/Logos/oranynth_logo.png.meta
