+++ Assets/_Project/Images/Logos/oranynth_logo.xcf.meta
