+++ Assets/_Project/Images/UI.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 44d89d28fc3c6454dbac9c1f1f7315ec
-PrefabImporter:
+guid: dcc1c356da4c06f468d6cb96e3a300a5
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
