+++ Assets/_Project/Images/UI/button_default.png.meta
@@ -1,10 +1,10 @@
 fileFormatVersion: 2
-guid: 177ee615af71dc74c8757558ba70a450
+guid: c386184965aa3da45af58a19cac0641a
 TextureImporter:
   internalIDToNameTable:
   - first:
-      213: -1706920167987475274
-    second: powerup_time_freeze_0
+      213: -2771146889623541776
+    second: button_default_0
   externalObjects: {}
   serializedVersion: 13
   mipmaps:
@@ -109,11 +109,24 @@ TextureImporter:
     ignorePlatformSupport: 0
     androidETC2FallbackOverride: 0
     forceMaximumCompressionQuality_BC6H_BC7: 0
+  - serializedVersion: 4
+    buildTarget: WebGL
+    maxTextureSize: 2048
+    resizeAlgorithm: 0
+    textureFormat: -1
+    textureCompression: 1
+    compressionQuality: 50
+    crunchedCompression: 0
+    allowsAlphaSplitting: 0
+    overridden: 0
+    ignorePlatformSupport: 0
+    androidETC2FallbackOverride: 0
+    forceMaximumCompressionQuality_BC6H_BC7: 0
   spriteSheet:
     serializedVersion: 2
     sprites:
     - serializedVersion: 2
-      name: powerup_time_freeze_0
+      name: button_default_0
       rect:
         serializedVersion: 2
         x: 0
@@ -121,15 +134,15 @@ TextureImporter:
         width: 500
         height: 500
       alignment: 0
-      pivot: {x: 0, y: 0}
-      border: {x: 0, y: 0, z: 0, w: 0}
+      pivot: {x: 0.5, y: 0.5}
+      border: {x: 20, y: 20, z: 20, w: 20}
       customData: 
       outline: []
       physicsShape: []
       tessellationDetail: -1
       bones: []
-      spriteID: 6bc082f962dcf48e0800000000000000
-      internalID: -1706920167987475274
+      spriteID: 0f76f870698ea89d0800000000000000
+      internalID: -2771146889623541776
       vertices: []
       indices: 
       edges: []
@@ -138,7 +151,7 @@ TextureImporter:
     customData: 
     physicsShape: []
     bones: []
-    spriteID: 
+    spriteID: 27e37b5929d16b148ab76b9ffe7346d5
     internalID: 0
     vertices: []
     indices: 
@@ -148,7 +161,7 @@ TextureImporter:
     spriteCustomMetadata:
       entries: []
     nameFileIdTable:
-      powerup_time_freeze_0: -1706920167987475274
+      button_default_0: -2771146889623541776
   mipmapLimitGroupName: 
   pSDRemoveMatte: 0
   userData:
