+++ Assets/_Project/Images/UI/button_sprites.xcf.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 44d89d28fc3c6454dbac9c1f1f7315ec
-PrefabImporter:
+guid: 293343e82762b6b459c55e81634b139f
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
