+++ Assets/_Project/Images/collectible_icons.xcf.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9eed47eb5feb5e948ad0adba85781582
+guid: f5a12091316bebd47a1b7c883a72e8f6
 DefaultImporter:
   externalObjects: {}
   userData:
