+++ Assets/_Project/Scenes/1 - Game.unity
@@ -673,9 +673,9 @@ MonoBehaviour:
   SpawnArea: {fileID: 329422966}
   NoSpawnArea: {fileID: 329422968}
   spawnableObjects:
-  - chanceOfSpawn: 0
-    prefab: {fileID: 7483760908730672677, guid: 669a89e21ca7d7d4fbfef4378440a88f, type: 3}
   - chanceOfSpawn: 100
+    prefab: {fileID: 7483760908730672677, guid: 669a89e21ca7d7d4fbfef4378440a88f, type: 3}
+  - chanceOfSpawn: 5
     prefab: {fileID: 4697666405567495450, guid: 44d89d28fc3c6454dbac9c1f1f7315ec, type: 3}
   StartCount: 20
   SpawnRate: 1
@@ -1614,7 +1614,7 @@ MonoBehaviour:
     m_SelectOnDown: {fileID: 0}
     m_SelectOnLeft: {fileID: 0}
     m_SelectOnRight: {fileID: 0}
-  m_Transition: 1
+  m_Transition: 2
   m_Colors:
     m_NormalColor: {r: 1, g: 1, b: 1, a: 1}
     m_HighlightedColor: {r: 0.9607843, g: 0.9607843, b: 0.9607843, a: 1}
@@ -1624,10 +1624,10 @@ MonoBehaviour:
     m_ColorMultiplier: 1
     m_FadeDuration: 0.1
   m_SpriteState:
-    m_HighlightedSprite: {fileID: 0}
-    m_PressedSprite: {fileID: 0}
-    m_SelectedSprite: {fileID: 0}
-    m_DisabledSprite: {fileID: 0}
+    m_HighlightedSprite: {fileID: 4200140477665312040, guid: 1cc456240477a714e80192b4212ad419, type: 3}
+    m_PressedSprite: {fileID: 4200140477665312040, guid: 1cc456240477a714e80192b4212ad419, type: 3}
+    m_SelectedSprite: {fileID: 4200140477665312040, guid: 1cc456240477a714e80192b4212ad419, type: 3}
+    m_DisabledSprite: {fileID: -2771146889623541776, guid: c386184965aa3da45af58a19cac0641a, type: 3}
   m_AnimationTriggers:
     m_NormalTrigger: Normal
     m_HighlightedTrigger: Highlighted
@@ -1671,7 +1671,7 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_Sprite: {fileID: 10905, guid: 0000000000000000f000000000000000, type: 0}
+  m_Sprite: {fileID: -2771146889623541776, guid: c386184965aa3da45af58a19cac0641a, type: 3}
   m_Type: 1
   m_PreserveAspect: 0
   m_FillCenter: 1
@@ -2902,7 +2902,7 @@ MonoBehaviour:
     m_SelectOnDown: {fileID: 0}
     m_SelectOnLeft: {fileID: 0}
     m_SelectOnRight: {fileID: 0}
-  m_Transition: 1
+  m_Transition: 2
   m_Colors:
     m_NormalColor: {r: 1, g: 1, b: 1, a: 1}
     m_HighlightedColor: {r: 0.9607843, g: 0.9607843, b: 0.9607843, a: 1}
@@ -2912,10 +2912,10 @@ MonoBehaviour:
     m_ColorMultiplier: 1
     m_FadeDuration: 0.1
   m_SpriteState:
-    m_HighlightedSprite: {fileID: 0}
-    m_PressedSprite: {fileID: 0}
-    m_SelectedSprite: {fileID: 0}
-    m_DisabledSprite: {fileID: 0}
+    m_HighlightedSprite: {fileID: 4200140477665312040, guid: 1cc456240477a714e80192b4212ad419, type: 3}
+    m_PressedSprite: {fileID: 4200140477665312040, guid: 1cc456240477a714e80192b4212ad419, type: 3}
+    m_SelectedSprite: {fileID: 4200140477665312040, guid: 1cc456240477a714e80192b4212ad419, type: 3}
+    m_DisabledSprite: {fileID: -2771146889623541776, guid: c386184965aa3da45af58a19cac0641a, type: 3}
   m_AnimationTriggers:
     m_NormalTrigger: Normal
     m_HighlightedTrigger: Highlighted
@@ -2959,7 +2959,7 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_Sprite: {fileID: 10905, guid: 0000000000000000f000000000000000, type: 0}
+  m_Sprite: {fileID: -2771146889623541776, guid: c386184965aa3da45af58a19cac0641a, type: 3}
   m_Type: 1
   m_PreserveAspect: 0
   m_FillCenter: 1
@@ -3408,7 +3408,7 @@ MonoBehaviour:
     m_SelectOnDown: {fileID: 0}
     m_SelectOnLeft: {fileID: 0}
     m_SelectOnRight: {fileID: 0}
-  m_Transition: 1
+  m_Transition: 2
   m_Colors:
     m_NormalColor: {r: 1, g: 1, b: 1, a: 1}
     m_HighlightedColor: {r: 0.9607843, g: 0.9607843, b: 0.9607843, a: 1}
@@ -3418,10 +3418,10 @@ MonoBehaviour:
     m_ColorMultiplier: 1
     m_FadeDuration: 0.1
   m_SpriteState:
-    m_HighlightedSprite: {fileID: 0}
-    m_PressedSprite: {fileID: 0}
-    m_SelectedSprite: {fileID: 0}
-    m_DisabledSprite: {fileID: 0}
+    m_HighlightedSprite: {fileID: 4200140477665312040, guid: 1cc456240477a714e80192b4212ad419, type: 3}
+    m_PressedSprite: {fileID: 4200140477665312040, guid: 1cc456240477a714e80192b4212ad419, type: 3}
+    m_SelectedSprite: {fileID: 4200140477665312040, guid: 1cc456240477a714e80192b4212ad419, type: 3}
+    m_DisabledSprite: {fileID: -2771146889623541776, guid: c386184965aa3da45af58a19cac0641a, type: 3}
   m_AnimationTriggers:
     m_NormalTrigger: Normal
     m_HighlightedTrigger: Highlighted
@@ -3465,7 +3465,7 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_Sprite: {fileID: 10905, guid: 0000000000000000f000000000000000, type: 0}
+  m_Sprite: {fileID: -2771146889623541776, guid: c386184965aa3da45af58a19cac0641a, type: 3}
   m_Type: 1
   m_PreserveAspect: 0
   m_FillCenter: 1
@@ -3541,7 +3541,7 @@ MonoBehaviour:
     m_SelectOnDown: {fileID: 0}
     m_SelectOnLeft: {fileID: 0}
     m_SelectOnRight: {fileID: 0}
-  m_Transition: 1
+  m_Transition: 2
   m_Colors:
     m_NormalColor: {r: 1, g: 1, b: 1, a: 1}
     m_HighlightedColor: {r: 0.9607843, g: 0.9607843, b: 0.9607843, a: 1}
@@ -3551,10 +3551,10 @@ MonoBehaviour:
     m_ColorMultiplier: 1
     m_FadeDuration: 0.1
   m_SpriteState:
-    m_HighlightedSprite: {fileID: 0}
-    m_PressedSprite: {fileID: 0}
-    m_SelectedSprite: {fileID: 0}
-    m_DisabledSprite: {fileID: 0}
+    m_HighlightedSprite: {fileID: 4200140477665312040, guid: 1cc456240477a714e80192b4212ad419, type: 3}
+    m_PressedSprite: {fileID: 4200140477665312040, guid: 1cc456240477a714e80192b4212ad419, type: 3}
+    m_SelectedSprite: {fileID: 4200140477665312040, guid: 1cc456240477a714e80192b4212ad419, type: 3}
+    m_DisabledSprite: {fileID: -2771146889623541776, guid: c386184965aa3da45af58a19cac0641a, type: 3}
   m_AnimationTriggers:
     m_NormalTrigger: Normal
     m_HighlightedTrigger: Highlighted
@@ -3598,7 +3598,7 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_Sprite: {fileID: 10905, guid: 0000000000000000f000000000000000, type: 0}
+  m_Sprite: {fileID: -2771146889623541776, guid: c386184965aa3da45af58a19cac0641a, type: 3}
   m_Type: 1
   m_PreserveAspect: 0
   m_FillCenter: 1
@@ -4261,7 +4261,7 @@ MonoBehaviour:
   TotalTime: 0
   CorrectCount: 0
   WrongCount: 0
-  freezeTime: 0
+  FreezeTime: 0
   Countdown: 0
   State: 0
 --- !u!114 &1820963859
@@ -4604,6 +4604,10 @@ MonoBehaviour:
   - {fileID: 1763933197}
   - {fileID: 1111340541}
   - {fileID: 1640744406}
+  PanelSelects:
+  - {fileID: 0}
+  - {fileID: 713946047}
+  - {fileID: 1272810233}
 --- !u!4 &2099303046
 Transform:
   m_ObjectHideFlags: 0
